pyproject.toml
[project]
name = "thesis-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "azure-search-documents>=11.6.0",
    "fastapi[standard]>=0.120.4",
    "langchain>=1.0.2",
    "langchain-core>=1.0.1",
    "langgraph>=1.0.1",
    "openai>=2.6.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
]
